Add SelfCleanWorkChain for all subworkflow
