+++ aiida_sssp_workflow/workflows/__init__.py
@@ -1,0 +1,37 @@
+from aiida.engine import WorkChain
+from aiida.orm import Bool, CalcJobNode
+
+
+class SelfCleanWorkChain(WorkChain):
+    @classmethod
+    def define(cls, spec):
+        super().define(spec)
+        spec.input(
+            "clean_workdir",
+            valid_type=Bool,
+            default=lambda: Bool(False),
+            help="If `True`, work directories of all called calculation will be cleaned at the end of execution.",
+        )
+
+    def on_terminated(self):
+        """Clean the working directories of all child calculations if `clean_workdir=True` in the inputs."""
+        super().on_terminated()
+
+        if self.inputs.clean_workdir.value is False:
+            self.report("remote folders will not be cleaned")
+            return
+
+        cleaned_calcs = []
+
+        for called_descendant in self.node.called_descendants:
+            if isinstance(called_descendant, CalcJobNode):
+                try:
+                    called_descendant.outputs.remote_folder._clean()  # pylint: disable=protected-access
+                    cleaned_calcs.append(called_descendant.pk)
+                except (IOError, OSError, KeyError):
+                    pass
+
+        if cleaned_calcs:
+            self.report(
+                f"cleaned remote folders of calculations: {' '.join(map(str, cleaned_calcs))}"
+            )
